+++ intelmq/tests/lib/test_upgrades.py
@@ -6,6 +6,24 @@ import unittest
 import intelmq.lib.upgrades as upgrades
 
 
+V202 = {"test-collector": {
+    "group": "Collector",
+    "module": "intelmq.bots.collectors.http.collector_http",
+    "parameters": {
+        "feed": "Feed"
+    }
+},
+}
+V202_EXP = {"test-collector": {
+    "group": "Collector",
+    "module": "intelmq.bots.collectors.http.collector_http",
+    "parameters": {
+        "name": "Feed"
+    }
+},
+}
+
+
 def generate_function(function):
     def test_function(self):
         """ Test if no errors happen for upgrade function %s. """ % function.__name__
@@ -23,7 +41,6 @@ class TestUpgradeLib(unittest.TestCase):
             self.mapping_list.extend((x for x in values))
             self.mapping_list_name.extend((x.__name__ for x in values))
 
-
     def test_all_functions_used(self):
         self.assertEqual(len(self.mapping_list_name),
                          len(set(self.mapping_list_name)),
@@ -35,6 +52,12 @@ class TestUpgradeLib(unittest.TestCase):
         self.assertEqual(set(self.allfs), set(self.modulefs),
                          msg='v* functions in the module do not '
                              'match functions in __all__.')
+
+    def test_v202_fixes(self):
+        """ Test v202_feed_name """
+        result = upgrades.v202_fixes({}, V202, False)
+        self.assertTrue(result[0])
+        self.assertEqual(V202_EXP, result[2])
 
 
 for name in upgrades.__all__: